+++ lmrj-modules/lmrj-module-rms/src/main/java/com/lmrj/rms/recipe/entity/RmsRecipe.java
@@ -59,7 +59,7 @@ public class RmsRecipe extends BaseDataEntity {
     /**配方版本号*/
     @TableField(value = "version_no")
     @Excel(name = "程序版本号", orderNum = "1", width = 20)
-    private Double versionNo;
+    private Integer versionNo;
     /**状态 0、创建（不可用）Y、生效（可用）N、停用（不可用）*/
     @TableField(value = "status")
     private String status;
